[BE] Seperate main controller
